+++ Game/MeadowRemixOptions.cs
@@ -43,8 +43,9 @@ public class RainMeadowOptions : OptionInterface
 
 
     public readonly Configurable<int> ArtificerParryDistance;
+    public readonly Configurable<int> ArtificerStunDistance;
     public readonly Configurable<bool> BlockMaul;
-    public readonly Configurable<bool> BlockArtiStun, ArenaAllowMidJoin;
+    public readonly Configurable<bool> ArenaAllowMidJoin;
     public readonly Configurable<bool> WearingCape;
     public readonly Configurable<bool> SlugpupHellBackground;
     public readonly Configurable<bool> StoryItemSteal;
@@ -170,9 +171,9 @@ public class RainMeadowOptions : OptionInterface
         PainCatThrows = config.Bind("PainCatThrows", false);
         PainCatEgg = config.Bind("PainCatEgg", true);
         PainCatLizard = config.Bind("PainCatLizard", true);
-        ArtificerParryDistance = config.Bind("ArtificerParryDistance", 90);
+        ArtificerParryDistance = config.Bind("ArtificerParryDistance", 100);
+        ArtificerStunDistance = config.Bind("ArtificerStunDistance", 100);
         BlockMaul = config.Bind("BlockMaul", false);
-        BlockArtiStun = config.Bind("BlockArtiStun", false);
         WeaverWatcher = config.Bind("WeaverWatcher", false);
         VoidMaster = config.Bind("VoidMaster", false);
         AmoebaDuration = config.Bind("AmoebaDuration", 7);