+++ Menu/Components/OnlineSlugcatAbilitiesInterface.cs
@@ -115,17 +115,17 @@ namespace RainMeadow.UI.Components
 
         public class MSCSettingsPage : SettingsPage, CheckBox.IOwnCheckBox
         {
-            public const string SAINOT = "SAINOT", PAINCATTHROWS = "PAINCATTHROWS", PAINCATEGG = "PAINCATEGG", DISABLEARTISTUN = "DISABLEARTISTUN", DISABLEMAUL = "DISABLEMAUL", PAINCATLIZARD = "PAINCATLIZARD", SAINTASENSIONTIMER = "SAINTASCENSIONTIMER";
+            public const string SAINOT = "SAINOT", PAINCATTHROWS = "PAINCATTHROWS", PAINCATEGG = "PAINCATEGG", DISABLEMAUL = "DISABLEMAUL", PAINCATLIZARD = "PAINCATLIZARD", SAINTASENSIONTIMER = "SAINTASCENSIONTIMER";
             public SimpleButton? backButton;
             public MenuTabWrapper tabWrapper;
             public OpTextBox saintAscendDurationTimerTextBox;
             public MenuLabel saintAscendanceTimerLabel;
 
 
-            public OpTextBox artiExplosionTextBox, artiParryDistanceTextBox;
-            public MenuLabel artiExplosionLabel, artiParryDistanceLabel;
+            public OpTextBox artiExplosionTextBox, artiParryDistanceTextBox, artiStunDistanceTextBox;
+            public MenuLabel artiExplosionLabel, artiParryDistanceLabel, artiStunDistanceLabel;
 
-            public RestorableCheckbox blockMaulCheckBox, blockArtiStunCheckBox, sainotCheckBox, painCatEggCheckBox, painCatThrowsCheckBox, painCatLizardCheckBox;
+            public RestorableCheckbox blockMaulCheckBox, sainotCheckBox, painCatEggCheckBox, painCatThrowsCheckBox, painCatLizardCheckBox;
             public override string Name => "MSC Settings";
             public MSCSettingsPage(Menu.Menu menu, MenuObject owner, Vector2 spacing, string painCatName, float textSpacing = 300) : base(menu, owner)
             {
@@ -150,7 +150,7 @@ namespace RainMeadow.UI.Components
                 artiParryDistanceTextBox = new(new Configurable<int>(RainMeadow.rainMeadowOptions.ArtificerParryDistance.Value), positioner - spacing * 1 + new Vector2(-7.5f, 0), 40)
                 {
                     alignment = FLabelAlignment.Center,
-                    description = Translate("How far Artificer can parry from. 0 disables it. Vanilla default: 300. Online default: 90."),
+                    description = Translate("How far Artificer can parry from. 0 disables it. Vanilla default: 300. Online default: 100."),
                     accept = OpTextBox.Accept.Int
 
                 };
@@ -163,7 +163,22 @@ namespace RainMeadow.UI.Components
                 artiParryDistanceLabel.label.alignment = FLabelAlignment.Left;
                 new PatchedUIelementWrapper(tabWrapper, artiParryDistanceTextBox);
 
-                blockArtiStunCheckBox = new(menu, this, this, positioner - spacing * 2, textSpacing, Translate("Disable Artificer Stun:"), DISABLEARTISTUN, false, Translate("Prevent Artificer from stunning other players"));
+                artiStunDistanceTextBox = new(new Configurable<int>(RainMeadow.rainMeadowOptions.ArtificerStunDistance.Value), positioner - spacing * 2 + new Vector2(-7.5f, 0), 40)
+                {
+                    alignment = FLabelAlignment.Center,
+                    description = Translate("How far Artificer can stun other players. 0 disables it. Vanilla default: 200. Online default: 100."),
+                    accept = OpTextBox.Accept.Int
+
+                };
+                artiStunDistanceTextBox.OnValueUpdate += (UIconfig config, string value, string lastValue) =>
+                {
+                    if (!RainMeadow.isArenaMode(out ArenaMode arena)) return;
+                    arena.artiStunDistance = artiStunDistanceTextBox.valueInt;
+                };
+                artiStunDistanceLabel = new(menu, this, Translate("Artificer Stun Range"), artiStunDistanceTextBox.pos + new Vector2(-textSpacing * 1.5f + 7.5f, 3), new(textSpacing, 20), false);
+                artiStunDistanceLabel.label.alignment = FLabelAlignment.Left;
+                new PatchedUIelementWrapper(tabWrapper, artiStunDistanceTextBox);
+
                 blockMaulCheckBox = new(menu, this, this, positioner - spacing * 3, textSpacing, Translate("Disable Mauling:"), DISABLEMAUL, false, Translate("Prevent Artificer and <PAINCATNAME> from mauling"));
                 sainotCheckBox = new(menu, this, this, positioner - spacing * 4, textSpacing, Translate("Sain't:"), SAINOT, false, Translate("Disable Saint ascendance ability, but allow it to throw spears"));
                 saintAscendDurationTimerTextBox = new(new Configurable<int>(RainMeadow.rainMeadowOptions.ArenaSaintAscendanceTimer.Value), positioner - spacing * 5 + new Vector2(-7.5f, 0), 40)
@@ -182,7 +197,7 @@ namespace RainMeadow.UI.Components
                 painCatEggCheckBox = new(menu, this, this, positioner - spacing * 6, 300, Translate("<PAINCATNAME> gets egg at 0 throw skill:"), PAINCATEGG, description: Translate("If <PAINCATNAME> spawns with 0 throw skill, also spawn with Eggzer0"));
                 painCatThrowsCheckBox = new(menu, this, this, positioner - spacing * 7, 300, Translate("<PAINCATNAME> can always throw spears:"), PAINCATTHROWS, description: Translate("Always allow <PAINCATNAME> to throw spears, even if throw skill is 0"));
                 painCatLizardCheckBox = new(menu, this, this, positioner - spacing * 8, 300, Translate("<PAINCATNAME> sometimes gets a friend:"), PAINCATLIZARD, description: Translate("Allow <PAINCATNAME> to rarely spawn with a little friend"));
-                this.SafeAddSubobjects(tabWrapper, blockMaulCheckBox, blockArtiStunCheckBox, sainotCheckBox, saintAscendanceTimerLabel, painCatEggCheckBox, painCatThrowsCheckBox, painCatLizardCheckBox, artiExplosionLabel, artiParryDistanceLabel);
+                this.SafeAddSubobjects(tabWrapper, blockMaulCheckBox, sainotCheckBox, saintAscendanceTimerLabel, painCatEggCheckBox, painCatThrowsCheckBox, painCatLizardCheckBox, artiExplosionLabel, artiParryDistanceLabel, artiStunDistanceLabel);
                 string Translate(string s) => menu.LongTranslate(s).Replace("<PAINCATNAME>", painCatName);
             }
             public void SyncMenuObjectStatus(MenuObject obj)
@@ -194,8 +209,8 @@ namespace RainMeadow.UI.Components
             {
                 MoreSlugcats.MoreSlugcats.cfgArtificerExplosionCapacity.Value = artiExplosionTextBox.valueInt;
                 RainMeadow.rainMeadowOptions.ArtificerParryDistance.Value = artiParryDistanceTextBox.valueInt;
+                RainMeadow.rainMeadowOptions.ArtificerStunDistance.Value = artiStunDistanceTextBox.valueInt;
                 RainMeadow.rainMeadowOptions.BlockMaul.Value = blockMaulCheckBox.Checked;
-                RainMeadow.rainMeadowOptions.BlockArtiStun.Value = blockArtiStunCheckBox.Checked;
                 RainMeadow.rainMeadowOptions.ArenaSAINOT.Value = sainotCheckBox.Checked;
                 RainMeadow.rainMeadowOptions.PainCatEgg.Value = painCatEggCheckBox.Checked;
                 RainMeadow.rainMeadowOptions.PainCatThrows.Value = painCatThrowsCheckBox.Checked;
@@ -210,6 +225,7 @@ namespace RainMeadow.UI.Components
                 arena.arenaSaintAscendanceTimer = saintAscendDurationTimerTextBox.valueInt;
                 arena.artiExplosionCount = artiExplosionTextBox.valueInt;
                 arena.artiParryDistance = artiParryDistanceTextBox.valueInt;
+                arena.artiStunDistance = artiStunDistanceTextBox.valueInt;
             }
             public override void SelectAndCreateBackButtons(SettingsPage? previousSettingPage, bool forceSelectedObject)
             {
@@ -238,6 +254,7 @@ namespace RainMeadow.UI.Components
                     ShowSyncInTextbox(saintAscendDurationTimerTextBox, greyoutAll, arena.arenaSaintAscendanceTimer);
                     ShowSyncInTextbox(artiExplosionTextBox, greyoutAll, arena.artiExplosionCount);
                     ShowSyncInTextbox(artiParryDistanceTextBox, greyoutAll, arena.artiParryDistance);
+                    ShowSyncInTextbox(artiStunDistanceTextBox, greyoutAll, arena.artiStunDistance);
                 }
 
             }
@@ -248,6 +265,7 @@ namespace RainMeadow.UI.Components
                 saintAscendanceTimerLabel.label.color = saintAscendDurationTimerTextBox.rect.colorEdge;
                 artiExplosionLabel.label.color = artiExplosionTextBox.rect.colorEdge;
                 artiParryDistanceLabel.label.color = artiParryDistanceTextBox.rect.colorEdge;
+                artiStunDistanceLabel.label.color = artiStunDistanceTextBox.rect.colorEdge;
 
             }
             public bool GetChecked(CheckBox box)
@@ -256,7 +274,6 @@ namespace RainMeadow.UI.Components
                 if (RainMeadow.isArenaMode(out ArenaMode arena))
                 {
                     if (id == DISABLEMAUL) return arena.disableMaul;
-                    if (id == DISABLEARTISTUN) return arena.disableArtiStun;
                     if (id == SAINOT) return arena.sainot;
                     if (id == PAINCATEGG) return arena.painCatEgg;
                     if (id == PAINCATTHROWS) return arena.painCatThrows;
@@ -269,7 +286,6 @@ namespace RainMeadow.UI.Components
                 if (!RainMeadow.isArenaMode(out ArenaMode arena)) return;
                 string id = box.IDString;
                 if (id == DISABLEMAUL) arena.disableMaul = c; //owner can only edit it, its fine
-                if (id == DISABLEARTISTUN) arena.disableArtiStun = c;
                 if (id == SAINOT) arena.sainot = c;
                 if (id == PAINCATEGG) arena.painCatEgg = c;
                 if (id == PAINCATTHROWS) arena.painCatThrows = c;